Isolate provider usage extraction
Move provider token usage parsing and accumulation out of runtime-loader.ts so chat, responses, Anthropic, and Google adapters share a focused helper module while keeping every provider-facing result shape unchanged. This is a move-only slice to continue shrinking the runtime-loader monolith without touching stream state machines or request builders.

Constraint: Behavior-preserving refactor only; no token field mapping or accumulation semantics changes

Rejected: Merge usage helpers with embedding response parsing | embedding helpers return embedding-specific token totals and would blur runtime result boundaries

Rejected: Split provider stream adapters in this pass | higher-risk state machines are still better left intact

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep provider usage mappings aligned with existing runtime-loader tests before expanding fields

Tested: VF_DISABLE_LRU_INTERVAL=1 deno test --no-check --allow-all src/provider/runtime-loader.test.ts src/provider/runtime-loader/provider-request-init.test.ts; deno fmt --check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-usage.ts; DENO_NO_PACKAGE_JSON=1 deno lint src/provider/runtime-loader.ts src/provider/runtime-loader/provider-usage.ts; deno check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-usage.ts; deno task typecheck

Not-tested: Full deno task test suite before push; pre-push hook will run the full unit suite
